fix - change in project code structure
